feat(scheduling): implement azure, cinder, ebs and gce as filter plugin
